added the ability to remove shapes
**READ THIS**
when making a shape, the method will return a number corresponding to
the shape. you can now use this integer to remove and remake that shape
at anytime. I would reccomend creating all possible shapes at the
beginning of the programs, and store an array of ints, and choose which
shapes to show or remove from those number